+++ crates/engine/src/vm/instruction/dml/table_update.rs
@@ -210,7 +210,7 @@ pub(crate) fn update_table(
 
 				let old_created_at =
 					txn.get(&row_key)?.expect("row must exist for update").row.created_at_nanos();
-				row.set_timestamps(old_created_at, services.runtime_context.clock.now_nanos() as u64);
+				row.set_timestamps(old_created_at, services.runtime_context.clock.now_nanos());
 
 				let stored_row = txn.update_table(table.clone(), row_number, row)?;
